+++ universal/index.d.ts
@@ -8,7 +8,8 @@ import {
 export {
 	Schema,
 	ParsedObjectsResult,
-	Error,
+	SchemaParseCellValueError,
+	CellValueRequiredError,
 	CellValue,
 	Row,
 	Integer,